Package description

Create a wide range of interactive, zoomable vector maps. This package is an R binding for the geoviz javascript library (https://github.com/riatelab/geoviz), itself based on the d3.js ecosystem ported by Mike Bostock (https://d3js.org/). Like the original javascript library, the package takes advantage of d3's many features: proportional symbols, pictograms, typologies, choropleth maps, Spikes, tiles, Dorling cartograms, etc. It can also be used to create pretty static vectorial maps in svg format, suitable for editorial cartography.

See also

Author

Maintainer: Nicolas Lambert nicolas.lambert@cnrs.fr (ORCID)

Authors:

Nicolas Lambert nicolas.lambert@cnrs.fr (ORCID)

Other contributors:

Timothée Giraud timothee.giraud@cnrs.fr (ORCID) [contributor]